Consolidating save.go and load.go into storage.go
