+++ internal/wallet/storage.go
@@ -1,20 +1,42 @@
 package wallet
 
 import (
-	"bufio"
-	"fmt"
+	"log"
 	"os"
 	"path/filepath"
-	"strings"
-	"time"
+
+	"bufio"
+	"fmt"
 
 	"github.com/joho/godotenv"
+
+	"strings"
+	"time"
 )
 
 const (
 	walletDir  = "./wallets"
 	timeFormat = "2006-01-02T15:04:05Z"
 )
+
+func saveWalletData(walletName, encryptedSeedPhrase, encryptedPubPass, encryptedPrivPass, encryptedBirthdate string) {
+	err := os.MkdirAll(walletDir, os.ModePerm)
+	if err != nil {
+		log.Fatalf("Error creating wallet directory: %v", err)
+	}
+
+	envFile := filepath.Join(walletDir, walletName+".env")
+	err = godotenv.Write(map[string]string{
+		"ENCRYPTED_SEED_PHRASE":        encryptedSeedPhrase,
+		"ENCRYPTED_PUBLIC_PASSPHRASE":  encryptedPubPass,
+		"ENCRYPTED_PRIVATE_PASSPHRASE": encryptedPrivPass,
+		"ENCRYPTED_BIRTHDATE":          encryptedBirthdate,
+	}, envFile)
+
+	if err != nil {
+		log.Fatalf("Error saving encrypted data: %v", err)
+	}
+}
 
 func loadWallet(walletName string) (string, string, string, time.Time, error) {
 	envFile := filepath.Join(walletDir, walletName+".env")
